[SWM-293] Feat : is_playlist && is_enrolled props in card ui
