+++ src/components/ui/lectureCard/HorizontalSmallLectureCard.tsx
@@ -1,10 +1,14 @@
 import Image from 'next/image';
+import ThumbnailBadge from '../badge/ThumbnailBadge';
 
 type Props = {
   src: string;
   alt: string;
   children: React.ReactNode;
+  isPlaylist?: boolean;
+  isEnrolled?: boolean;
   border?: boolean;
+  disabled?: boolean;
   onClick?: () => void;
   onMouseOver?: () => void;
 };
@@ -13,15 +17,18 @@ export default function HorizontalSmallLectureCard({
   src,
   alt,
   children,
+  isPlaylist = false,
+  isEnrolled = false,
   border = true,
+  disabled = false,
   onClick,
   onMouseOver
 }: Props) {
   return (
     <div
-      className={`relative flex w-full gap-4 p-3 whitespace-normal h-min-[11rem] bg-sroom-white text-sroom-black-400 ${
-        border ? 'border border-sroom-gray-400' : ''
-      }`}
+      className={`relative flex w-full gap-4 p-3 whitespace-normal h-[11rem] text-sroom-black-400 ${
+        disabled ? 'bg-sroom-gray-200' : 'bg-sroom-white'
+      } ${border ? 'border border-sroom-gray-400' : ''}`}
     >
       <div className='w-full max-w-[15rem] flex items-center'>
         <div className='relative w-full h-0 pb-[56.25%]'>
@@ -36,6 +43,23 @@ export default function HorizontalSmallLectureCard({
               src={src}
               alt={alt}
             />
+            {isEnrolled && (
+              <div className='absolute top-0 left-0'>
+                <ThumbnailBadge
+                  title='수강 중'
+                  className='bg-sroom-black-400'
+                />
+              </div>
+            )}
+            {isPlaylist && (
+              <div
+                className={`absolute top-0 ${
+                  isEnrolled ? 'left-[4rem]' : 'left-0'
+                }`}
+              >
+                <ThumbnailBadge title='재생목록' className='bg-sroom-brand' />
+              </div>
+            )}
           </div>
         </div>
       </div>
